refactor: DogStatistics 컴포넌트 생성 and JounalsState 생성
